add swal for delete confirmation

+++ src/Random component/Mytoy.jsx
@@ -2,6 +2,7 @@ import React, { useContext, useEffect, useState } from 'react';
 import { ContextAuth } from '../Routes/AuthenticationCenter';
 import { Table } from 'react-bootstrap';
 import { Link } from 'react-router-dom';
+import Swal from 'sweetalert2';
 
 const Mytoy = () => {
 let {user}=useContext(ContextAuth)
@@ -22,13 +23,38 @@ fetch(url)
 
 let deletetoy=id=>{
 
-fetch(`https://b7a11-toy-marketplace-server-three.vercel.app/alltoy/${id}`, {
-  method:"DELETE",
-}) .then(res=> res.json())
-.then(data=>{
-console.log(data);
+  Swal.fire({
+    title: 'Are you sure?',
+    text: "You won't be able to revert this!",
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#3085d6',
+    cancelButtonColor: '#d33',
+    confirmButtonText: 'Yes, delete it!'
+  }).then((result) => {
+    if (result.isConfirmed) {
+      fetch(`https://b7a11-toy-marketplace-server-three.vercel.app/alltoy/${id}`, {
+        method:"DELETE",
+      }) .then(res=> res.json())
+      .then(data=>{
+      console.log(data);
+      if (data.deletedCount>0) {
+        Swal.fire(
+           'Deleted!',
+           'Your file has been deleted.',
+           'success'
+         )
+   
+      
+       }
+      })
+   
+}
+  })
 
-})
+  
+
+
 
 }
